Initial public release of tabular
A 2-D grid model for holding "cells" geared for creating tables, for
various output models.  Current renderers for HTML, CSV and for
pretty-printed UTF-8 tables for fixed-character-cell displays (such as
the Unix visual-TTY terminal model).

Brief coverage in README.md, details of structure in Overview.md, there
are godocs, some tests (enough to catch major issues but not great) and
things work.  It's feature complete for a v1, this is not a barebones
MVP.  There's more that could be done for a v2, but want more real-world
feedback first.

Bumped API to 1.0.

There will be a new commit in a bit, setting up automated CI and adding
badges into the README.

Squashed history; original dev repo started 2016-10-03 and was worked on
intermittently between then and now.