wip: skeletal testing
Column alignment works "first time" that it's tested.  Promising!

Realized may need to return offsets larger than the width of a given
cell, where we align on a period which is not found, so that "42" aligns
above "142.3" such that summing the columns would be "184.3".

Not willing to do that tonight.