wip: more tests, isolated the problem
This algorithm works for _column_ alignment, where every cell in the
column uses the same alignment.  When overriding on a per-cell basis,
this falls apart.

Eg, everything by default is left-aligned, so the .toLeft is 0, except
for the one cell which is center/right aligned, so ... the center/right
gets aligned with the alignment point of the other cells, which is the
far left.

So broken algorithm.