Fix fixme tests which were flagged as bad-value
For some tests, we had issues with length measuring and, if memory
serves, a dependency doing something "off".  Today, the tests fail
because the "have" values were the values I'd marked in comments as what
we _really_ wanted, as opposed to what made the tests pass before.

So switch the tests to be the real desired values, remove the FIXME, and
let the tests pass once more.