Bug-fix property removal
The chaining logic was checking for "no child? oh, return" _before_
checking the current key, so would never remove an item which was the
last in the chain.

Renaming the child entry to be `child` not `propertySet` (anonymous
embed to named) made it much clearer.  Changing the child printing in
`GoString()` from `#v` to `#+v` also helped.

Fixed the removal logic.  Use the JSON skipability tests to exercise
this logic.

Since we return `nil` for "not found", I decided that there's no point
to being able to store a nil value, so have changed the API definition
such that trying to store nil _means_ removal.  (I did have a
`RemoveProperty` sentinel in here, but nil is actually clearer, since we
want `GetProperty()` to return just a value, no error signalling.)