Bump go-runewidth dependency to be more current
A lot of display width glitches can come from Go modules defaulting to
the oldest allowed version of a dependency and not picking up fixes; we
want better go-runewidth data, so force it.

The impact of this is severe enough that it warrants a new version tag
just for increasing this dependency, so that callers can specify a
tabular known to have up-to-date rune widths.