ci: go stable, action versions, fix persist-creds
Ran `zizmor .` to report issues.  We do not need subsequent git
actions to work against the repo, so do not need to persist credentials.

Bump action versions while I'm here.

Switch Go versions to be `stable` and `oldstable`.

Add a `workflow_dispatch` trigger, so we can test on newer versions of Go
without a new commit when the repo has been idle for a long time.